refactor: 결제 시간 수정
